ai-daily-brief/requirements.txt
feedparser==6.0.11
firecrawl-py==1.6.1
google-genai==0.1.0
openai==1.58.1
python-dateutil==2.9.0
